Move CDK invocation to its own package
