+++ java/clients/src/main/java/sleeper/clients/admin/AdminConfigStore.java
@@ -20,8 +20,8 @@ import com.amazonaws.services.s3.model.AmazonS3Exception;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import sleeper.clients.deploy.CdkDeploy;
-import sleeper.clients.deploy.InvokeCdkForInstance;
+import sleeper.clients.cdk.CdkDeploy;
+import sleeper.clients.cdk.InvokeCdkForInstance;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.UserDefinedInstanceProperty;
 import sleeper.configuration.properties.local.SaveLocalProperties;
